adjust styling, add script skeleton
I've spent a lot of time thinking about how I want this whole thing to
be structured which is why it is taking so long, but I think I now have
a basic outline that I can work with for the script which will come to
define the project. The question is whether or not I can write it in a
way that works and makes sense. Wish me luck!