Add logic for gameEndCheck script, adjust h2 class
So I have a (hopefully) working version of the logic I need to check if
the move that the current player just made is a game ending move (win or
draw). This works by first generating an array that contains the indexes
of where the player's respective marks are on the board ("X" marks for
the X player, "O" marks for the O player). Using this array, I check to
see if the pattern of indexes matches any of the patterns in the
winningCombinations array. If there is a match, then we have a winner!
if there is no match, we then need to check if the game is still going
on or not. I have implemented this by simply checking to see if there
are any spaces left on the board. If there are no spaces, then it must
be a draw. If there are still spaces left, there is no winner, and the
game can continue. All I need to do now is figure out how to implement
it properly.

I also added a class to the H2 in anticipation of having to target it
in my render function to update the UI when the player's turn changes.

I also added an object that contains the player's names and their
respective markers. I might update the code to use the markers instead
of just passing the playerTurn variable, but I am again anticipating
that I will need to have the players names stored somewhere to render
to the UI in order to allow players to input their own custom names.