Adjust code for accessbility, fix script issues
I added aria labels for each button so that a screen reader will
actually be able to read which position the button is in, since by
default this was obviously not the case. I also adjusted the styling so
that the intro screen is displayed at the top of the page, which should
make it fit and look better. Finally I fixed some small scripting
issues, particularly with placement of certain variables and function
calls. For example, the game logic would switch turns when a draw was
declared, which is unnecessary.